feat(kanban): 멤버 권한에 따른 인가 처리

+++ src/main/java/com/trello25/domain/kanban/service/KanbanService.java
@@ -2,7 +2,10 @@ package com.trello25.domain.kanban.service;
 
 import static com.trello25.exception.ErrorCode.BOARD_NOT_FOUND;
 import static com.trello25.exception.ErrorCode.KANBAN_NOT_FOUND;
+import static com.trello25.exception.ErrorCode.MEMBER_NOT_FOUND;
+import static com.trello25.exception.ErrorCode.UNAUTHORIZED_ACCESS;
 
+import com.trello25.domain.auth.dto.AuthUser;
 import com.trello25.domain.board.entity.Board;
 import com.trello25.domain.board.repository.BoardRepository;
 import com.trello25.domain.card.dto.response.CardResponse;
@@ -16,6 +19,9 @@ import com.trello25.domain.kanban.entity.Kanban;
 import com.trello25.domain.kanban.repository.KanbanRepository;
 import com.trello25.domain.kanbanposition.entity.KanbanPosition;
 import com.trello25.domain.kanbanposition.service.KanbanPositionService;
+import com.trello25.domain.member.entity.Member;
+import com.trello25.domain.member.entity.Permission;
+import com.trello25.domain.member.repository.MemberRepository;
 import com.trello25.exception.ApplicationException;
 import java.util.Comparator;
 import java.util.HashMap;
@@ -34,9 +40,14 @@ public class KanbanService {
     private final KanbanRepository kanbanRepository;
     private final KanbanPositionService kanbanPositionService;
     private final CardRepository cardRepository;
+    private final MemberRepository memberRepository;
 
     public void createKanban(AuthUser authUser, CreateKanbanRequest request) {
-        // TODO: 칸반 생성 권한을 가지고 있는 멤버인지 확인 필요, 로그인 기능 구현 완료 시 수정예정
+        Member member = memberRepository.findMemberForKanbanByBoardId(authUser.getId(), request.getBoardId())
+                .orElseThrow(() -> new ApplicationException(MEMBER_NOT_FOUND));
+        if (member.getPermission() == Permission.READ_ONLY) {
+            throw new ApplicationException(UNAUTHORIZED_ACCESS);
+        }
 
         Board board = boardRepository.findById(request.getBoardId())
                 .orElseThrow(() -> new ApplicationException(BOARD_NOT_FOUND));
@@ -49,28 +60,43 @@ public class KanbanService {
         kanbanPositionService.addKanban(kanban);
     }
 
-    public void deleteKanban(AuthUser authUser, long id) {
-        // TODO: 칸반 삭제 권한을 가지고 있는 멤버인지 확인 필요, 로그인 기능 구현 완료 시 수정예정
-
-        Kanban kanban = kanbanRepository.findByIdAndStatus(id, EntityStatus.ACTIVATED)
+    public void deleteKanban(AuthUser authUser, long kanbanId) {
+        Kanban kanban = kanbanRepository.findByIdAndStatus(kanbanId, EntityStatus.ACTIVATED)
                 .orElseThrow(() -> new ApplicationException(KANBAN_NOT_FOUND));
+
+        Member member = memberRepository.findMemberForKanbanByKanbanId(authUser.getId(), kanbanId)
+                .orElseThrow(() -> new ApplicationException(MEMBER_NOT_FOUND));
+        if (member.getPermission() == Permission.READ_ONLY) {
+            throw new ApplicationException(UNAUTHORIZED_ACCESS);
+        }
+
         kanban.delete();
         kanbanPositionService.deleteKanban(kanban);
     }
 
-    public void updateKanbanTitle(AuthUser authUser, long id, UpdateKanbanTitleRequest request) {
-        // TODO: 칸반 수정 권한을 가지고 있는 멤버인지 확인 필요, 로그인 기능 구현 완료 시 수정예정
-
-        Kanban kanban = kanbanRepository.findByIdAndStatus(id, EntityStatus.ACTIVATED)
+    public void updateKanbanTitle(AuthUser authUser, long kanbanId, UpdateKanbanTitleRequest request) {
+        Kanban kanban = kanbanRepository.findByIdAndStatus(kanbanId, EntityStatus.ACTIVATED)
                 .orElseThrow(() -> new ApplicationException(KANBAN_NOT_FOUND));
+
+        Member member = memberRepository.findMemberForKanbanByKanbanId(authUser.getId(), kanbanId)
+                .orElseThrow(() -> new ApplicationException(MEMBER_NOT_FOUND));
+        if (member.getPermission() == Permission.READ_ONLY) {
+            throw new ApplicationException(UNAUTHORIZED_ACCESS);
+        }
+
         kanban.updateTitle(request.getTitle());
     }
 
-    public void updateKanbanPosition(AuthUser authUser, long id, UpdateKanbanPositionRequest request) {
-        // TODO: 칸반 수정 권한을 가지고 있는 멤버인지 확인 필요, 로그인 기능 구현 완료 시 수정예정
-
-        Kanban kanban = kanbanRepository.findByIdAndStatus(id, EntityStatus.ACTIVATED)
+    public void updateKanbanPosition(AuthUser authUser, long kanbanId, UpdateKanbanPositionRequest request) {
+        Kanban kanban = kanbanRepository.findByIdAndStatus(kanbanId, EntityStatus.ACTIVATED)
                 .orElseThrow(() -> new ApplicationException(KANBAN_NOT_FOUND));
+
+        Member member = memberRepository.findMemberForKanbanByKanbanId(authUser.getId(), kanbanId)
+                .orElseThrow(() -> new ApplicationException(MEMBER_NOT_FOUND));
+        if (member.getPermission() == Permission.READ_ONLY) {
+            throw new ApplicationException(UNAUTHORIZED_ACCESS);
+        }
+
         kanbanPositionService.updateKanbanPosition(kanban, request.getPosition());
     }
